Deduplicate some GPU plugin definition code.
The `jaxlib/cuda_plugin_extension.cc` and `jaxlib/rocm_plugin_extension.cc` files were nearly identical so this change consolidates the shared implementation into a single target.

PiperOrigin-RevId: 704785926
