+++ jaxlib/gpu_plugin_extension.h
@@ -1,0 +1,27 @@
+/* Copyright 2024 The JAX Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+==============================================================================*/
+
+#ifndef JAXLIB_GPU_PLUGIN_EXTENSION_H_
+#define JAXLIB_GPU_PLUGIN_EXTENSION_H_
+
+#include "nanobind/nanobind.h"
+
+namespace xla {
+
+void BuildGpuPluginExtension(nanobind::module_& m);
+
+}  // namespace xla
+
+#endif  // JAXLIB_GPU_PLUGIN_EXTENSION_H_
